Update extensions.css
This assimilates an update by SW1FT to properly limit the border in about:addons to the #view-port and not the .main-content since that would also enclose the #header in Firefox 40+.  This change has additional code to avoid the change done to inContentUI.css by SW1FT in his solution.  I took extra caution to shift the #categories down only for Firefox 40+ to keep the dimensions the same as in previous versions.  This update also has a minor adjustment to some buttons in the #header (e.g., the button with the text "Some extensions could not be verified") to avoid a potential shift in the top of the #view-port.