Update panel.css
This update is needed for Firefox 41. It sets a border color value so that the elements in the Site Information Panel (a.k.a. Identity PopUp) can be delimited properly. The variable used was not set elsewhere in SimpleWhite. The update also overrides extra padding and margin settings that are set in browser.css (actually, browser-common.css) and used in Firefox 40 and earlier versions.  This way, incompatibility with earlier versions is avoided. 
Note: Using "important!" is avoided by simply increasing the priority of a selector, and avoiding the problems that stem from using it is more than "imporant!".