Update browser-common.css
This merges in the code from browser.css used by Aurora 42.0a2 at some moment.  By doing so, it incorporates the changes done by Lootyhoof to fix the problems with the display of the minimize, maximize, and close on Windows 10, and to remove vertical toolbar borders also on Windows 10. (But I have not tested it!)  It has some particular code changes done by Lootyhoof.  It brings many other changes which include the introduction of references to icons and readjustment the size of icons from add-ons.