Update browser-common.css
This now makes Firefox v.42 and later use the tracking protection icons that were already in the theme before introducing the changes regarding that version of Firefox.  Also the vertical size of those icons is corrected for those versions of Firefox.  This makes use of a better svg filter to darken properly some of the icons in the identity box when the security information panel becomes visible in those versions of Firefox.