Update browser-common.css
This now makes Firefox v.42 and later use the tracking protection icons that were already in the theme before the last change regarding that.  It also fixes the vertical size of those icons in those versions of Firefox.  This makes use of a better svg filter to darken properly some of the icons in the identity box when the security information panel is displayed.