Update preferences.css
This is to address issue #2 in @lychichem's fork.  This change probably should be matched with removing the corresponding code from search.css in the same folder (after some checking).  The code came form the newer Nighly 43.0a1 and has some small additions to make the columns align better (yet not perfectly!).