Update common.css
Merged in the changes from ostarget/general/global/in-content/common.css to make the about:preferences page look like the about:addons page and to make the about:privatebrowsing use black/gray instead of bright colors.  It this includes a correction by @lychichem.  This should make the about:preferences page work fine in the Apple implementation.