Update animationinspector.css
While working towards solving #7, I discovered the Developer Tools ~ Inspector ~ Animations was had problems with Firefox 40+.
* The animations were disabled at the end of running (a behavior prescribed for earlier versions).
* The speed factor had the incorrect color (black) when using the dark theme.
I substituted the css, which was verbatim from the default theme in Firefox 39, for the one used in the default for Firefox 41 (and 42!).  Then I made corrections so that the disabling behavior remains for Firefox 38-39 and to avoid a side effect on Firefox 37 from setting the height of the #players area.  Changes are still needed to support Firefox 43.0a2 properly.