Update animationinspector.css
This is a removes some of the additions in the previous commit making the code as if it was based on that from the default theme in Firefox 40 instead of Firefox 41.  The reason is that the code in Firefox 41 included a section on the "animation timeline component" (.animation-timeline) that appears not used until Firefox 43.  But this section of code is changed in the css of Firefox 43, and that is why this code was not able to fully support that version.  So this commit eliminates any partial support for Firefox 43, but leaves the code clean for a proper merge with the newer code.  This is done by removing all related to the "animation timeline component."