Update animationinspector.css
This change adds compatibility with Firefox 43.0a2.  This is done by merging in all the code related to the "animation timeline component" and the new toolbars in the css from the default theme in Firefox 43.0a2.  This change also assimilates the use of variables and the different HiDPI limit in the newer default css. The result is almost identical to the css in Firefox 43.0a2, but it retains code to provide compatibility with all previous versions back to Firefox 37.  This code has a reference to a new icon, animation-fast-track.svg, which has to be added in this theme.