Update newTab.css
This update (re)sets compatibility with rtl locales to the about:newtab page.  It has improved versions for the icons (SVG) in the search bar in that page.  The code was tested with to work properly in Arabic versions of Firefox 42.0b6 and Firefox 38.3.0 ESR.  It was tested also in English versions of Firefox 37 through 44.