Update conn-degraded.svg
This change makes the padlock for non-secure connections appear open.  A locked gray padlock was already used for connections with only domain validation in the address bar, and they are secure.  Although the green padlock is used for this in the default theme of Firefox 42 and later, the gray padlock stays in Simple White X.  Showing a plain locked padlock gives the wrong impression, regardless of it being gray.