Update broser.css
This update adopts the revised connection icons (#page-proxy-favicon) from the default theme for Firefox 42 and later, but leaves a gray padlock for connections with domain validation only.  It restores showing those icons in Firefox 45.0a1 (issue #15).  It changes the animation for showing the tracking protection icon.  It also adds a border to the bars that may show at the top and bottom of panel that lists the matches to the entry in address bar in Firefox 44.0a2 and 45.0a1.