Update newTab.css
This change revises the vertical margins among elements in about:newtab.  This makes the page look more cosistent between different versions of Firefox.  The change also adds a small margin to sides of the search bar so it does not touch the page borders when the window is too narrow.