Update browser-common.css
This change contains adjustments for the opacity of the identity icons.  With this change the developer toolbar (gcli) is no longer fixed to the dark theme, now it will show the light theme if selected in the developer tools.  Also, this change adds rules for the new usercontext icons.  There are some other additions for compatibility with Firefox 45.0a1.