Update commandline.css
This change makes the pop-up associated to the developer toolbar (gcli) adjust to the theme selected for the developer tools, dark or light, for versions of Firefox 40 and later.  For earlier versions of Firefox it will use a minor variation of the light theme.  This change also corrects a bug in Firefox that makes the pop-up be enclosed in a small window with scrollbars.