Update browser-common.css
This sets the use of special buttons when personas with bright buttons are used in 'Windows 10'.  It makes many other changes for the use of light-weight themes, a.k.a. personas.  It modifies the appearance of some buttons in some special situations.  It removes the rules for 'Classic Theme Restorer'.