Updated broswer/devtools/*.css
This makes the style for the Developer Tools match changes in the default theme in Firefox 47.  The frameless Inspector tools are now supported by this theme.  The developer tools are shown in a flatter style.  The light theme is more white. But here the toolbars background and borders will match the tones in the theme's browser toolbar.  There are a few more changes from the default theme.  For instance, the highlight background color for the editor with the dark theme is darker than in the default theme.