Update common.css
This change introduces the use of color variables that are needed for other style sheets.  It adds rules for about:debugging, so it looks like other about pages in the theme.  It has some other changes that affect various particular details in various pages, e.g., it makes the background color of the network error pages a bit lighter.