Update aboutNetworking.css
This change adds important code so about:networking is usable in Firefox 46 and later.  For such versions it will use a style similar to the one of about:addons, but no icons are available for the section tabs and the format within the sections is different due to its content.  In Firefox 45 and earlier the buttons in this page are changed (e.g., they will not use blue highlights).