Update *.css
Update the developer tools style sheets to use the new icons from Firefox 52 for the Light and Dark themes.  Revert the style of the Developer Tools' main toolbar to the style in Firefox 45 (for the Light theme) but minimize the contrast of the shade on the top of the tool selection buttons.  Change the Performance Tool view selection buttons for the Light and Dark themes to match the buttons in the Web Console Tool.  Make updates for compatibility with Firefox 52.  Improve the Firebug theme for the Developer Tools.